Moved ancillary types from `dataset.rs` Moved non-core Dataset methods into respective `raster` or `vector` modules.
