+++ src/gdal_major_object.rs
@@ -2,5 +2,5 @@ use gdal_sys::GDALMajorObjectH;
 
 /// Common trait for GDAL data types backed by [`GDALMajorObjectH`].
 pub trait MajorObject {
-    unsafe fn gdal_object_ptr(&self) -> GDALMajorObjectH;
+    fn gdal_object_ptr(&self) -> GDALMajorObjectH;
 }
