+++ src/vector/sql.rs
@@ -1,11 +1,15 @@
+use std::ffi::{CStr, CString};
 use std::ops::{Deref, DerefMut};
 
-use gdal_sys::GDALDatasetH;
+use crate::Dataset;
+use gdal_sys::{CPLErr, GDALDatasetH, OGRGeometryH};
 
-use crate::vector::{Layer, LayerAccess};
+use crate::errors::*;
+use crate::utils::_last_cpl_err;
+use crate::vector::{sql, Geometry, Layer, LayerAccess};
 
 /// The result of a SQL query executed by
-/// [`Dataset::execute_sql()`](crate::Dataset::execute_sql()). It is just a thin wrapper around a
+/// [`Dataset::execute_sql()`](Dataset::execute_sql()). It is just a thin wrapper around a
 /// [`Layer`], and you can treat it as such.
 #[derive(Debug)]
 pub struct ResultSet<'a> {
@@ -51,6 +55,108 @@ pub enum Dialect {
 
 pub(crate) const OGRSQL: &[u8] = b"OGRSQL\0";
 pub(crate) const SQLITE: &[u8] = b"SQLITE\0";
+
+/// [Dataset] methods relating to SQL over vector datasets.
+impl Dataset {
+    /// Execute a SQL query against the Dataset. It is equivalent to calling
+    /// [`GDALDatasetExecuteSQL`](https://gdal.org/api/raster_c_api.html#_CPPv421GDALDatasetExecuteSQL12GDALDatasetHPKc12OGRGeometryHPKc).
+    /// Returns a [`ResultSet`], which can be treated just as any other [`Layer`].
+    ///
+    /// Queries such as `ALTER TABLE`, `CREATE INDEX`, etc. have no [`ResultSet`], and return
+    /// `None`, which is distinct from an empty [`ResultSet`].
+    ///
+    /// # Arguments
+    /// * `query`: The SQL query
+    /// * `spatial_filter`: Limit results of the query to features that intersect the given
+    ///   [`Geometry`]
+    /// * `dialect`: The dialect of SQL to use. See
+    ///   <https://gdal.org/user/ogr_sql_sqlite_dialect.html>
+    ///
+    /// # Example
+    ///
+    /// ```
+    /// # use gdal::Dataset;
+    /// # use std::path::Path;
+    /// use gdal::vector::sql;
+    /// use gdal::vector::LayerAccess;
+    ///
+    /// let ds = Dataset::open(Path::new("fixtures/roads.geojson")).unwrap();
+    /// let query = "SELECT kind, is_bridge, highway FROM roads WHERE highway = 'pedestrian'";
+    /// let mut result_set = ds.execute_sql(query, None, sql::Dialect::DEFAULT).unwrap().unwrap();
+    ///
+    /// assert_eq!(10, result_set.feature_count());
+    ///
+    /// for feature in result_set.features() {
+    ///     let highway = feature
+    ///         .field("highway")
+    ///         .unwrap()
+    ///         .unwrap()
+    ///         .into_string()
+    ///         .unwrap();
+    ///
+    ///     assert_eq!("pedestrian", highway);
+    /// }
+    /// ```
+    pub fn execute_sql<S: AsRef<str>>(
+        &self,
+        query: S,
+        spatial_filter: Option<&Geometry>,
+        dialect: Dialect,
+    ) -> Result<Option<ResultSet>> {
+        let query = CString::new(query.as_ref())?;
+
+        let dialect_c_str = match dialect {
+            Dialect::DEFAULT => None,
+            Dialect::OGR => Some(unsafe { CStr::from_bytes_with_nul_unchecked(OGRSQL) }),
+            Dialect::SQLITE => Some(unsafe { CStr::from_bytes_with_nul_unchecked(SQLITE) }),
+        };
+
+        self._execute_sql(query, spatial_filter, dialect_c_str)
+    }
+
+    fn _execute_sql(
+        &self,
+        query: CString,
+        spatial_filter: Option<&Geometry>,
+        dialect_c_str: Option<&CStr>,
+    ) -> Result<Option<sql::ResultSet>> {
+        let mut filter_geom: OGRGeometryH = std::ptr::null_mut();
+
+        let dialect_ptr = match dialect_c_str {
+            None => std::ptr::null(),
+            Some(d) => d.as_ptr(),
+        };
+
+        if let Some(spatial_filter) = spatial_filter {
+            filter_geom = unsafe { spatial_filter.c_geometry() };
+        }
+
+        let c_dataset = self.c_dataset();
+
+        unsafe { gdal_sys::CPLErrorReset() };
+
+        let c_layer = unsafe {
+            gdal_sys::GDALDatasetExecuteSQL(c_dataset, query.as_ptr(), filter_geom, dialect_ptr)
+        };
+
+        let cpl_err = unsafe { gdal_sys::CPLGetLastErrorType() };
+
+        if cpl_err != CPLErr::CE_None {
+            return Err(_last_cpl_err(cpl_err));
+        }
+
+        if c_layer.is_null() {
+            return Ok(None);
+        }
+
+        let layer = unsafe { Layer::from_c_layer(self, c_layer) };
+
+        Ok(Some(sql::ResultSet {
+            layer,
+            dataset: c_dataset,
+        }))
+    }
+}
 
 #[cfg(test)]
 mod tests {
